(svn r3008) [ 1247535 ] Native Support for Win64 (compile&run only) (michi_cc) Fix warning in graph_gui.c with const problem

+++ viewport.c
@@ -2229,7 +2229,7 @@ void SetObjectToPlace(CursorID icon, byte mode, WindowClass window_class, Window
 		VpStartPreSizing();
 
 	if ( (int)icon < 0)
-		SetAnimatedMouseCursor(_animcursors[~icon]);
+		SetAnimatedMouseCursor(_animcursors[~(int32)icon]);
 	else
 		SetMouseCursor(icon);
 }
